Add markdown support for posts
